Move embedding and extension functions to manager classes for easier maintenence
- fixes embeddings
- moves embeddings to embedding_manager file and uses class as mixin
- moves extensions to extension_manager file and uses class as mixin
- bumps version
